add protos and serializers

+++ firebase_firestore/lib/src/firebase/firestore/model/value/integer_value.dart
@@ -3,17 +3,18 @@
 // on 17/09/2018
 
 import 'package:firebase_firestore/src/firebase/firestore/model/value/number_value.dart';
+import 'package:fixnum/fixnum.dart';
 
 /// A wrapper for integer/long values in Firestore.
 class IntegerValue extends NumberValue {
-  final int _value;
+  final Int64 _value;
 
   const IntegerValue(this._value);
 
-  factory IntegerValue.valueOf(int value) => IntegerValue(value);
+  factory IntegerValue.valueOf(Int64 value) => IntegerValue(value);
 
   @override
-  int get value => _value;
+  Int64 get value => _value;
 
   @override
   bool operator ==(Object other) =>
